row set minimum

+++ src/tree/updater_quantile_hist.cc
@@ -792,8 +792,8 @@ template <bool default_left>
 inline std::pair<size_t, size_t> PartitionDenseKernel(
     const uint32_t *rid, const uint8_t *idx,
     const uint32_t offset, const int32_t split_cond,
-    const size_t istart, const size_t iend, uint32_t *p_left,
-    uint32_t *p_right) {
+    const size_t istart, const size_t iend, size_t *p_left,
+    size_t *p_right) {
   size_t ileft = 0;
   size_t iright = 0;
 
@@ -821,7 +821,7 @@ inline std::pair<size_t, size_t> PartitionDenseKernel(
 template<bool default_left>
 inline std::pair<size_t, size_t> PartitionSparseKernel(const uint32_t* rid,
     const uint8_t* idx, const uint32_t offset, const int32_t split_cond,
-    const size_t istart, const size_t iend, uint32_t* p_left, uint32_t* p_right,
+    const size_t istart, const size_t iend, size_t* p_left, size_t* p_right,
     bst_uint lower_bound, bst_uint upper_bound, const Column& column) {
 
   size_t ileft = 0;
@@ -880,8 +880,8 @@ void QuantileHistMaker::Builder::PartitionKernel(
     const int32_t split_cond, const ColumnMatrix& column_matrix,
     const GHistIndexMatrix& gmat, const RegTree& tree) {
   const uint32_t* rid = row_set_collection_[nid].begin;
-  uint32_t* p_left = partition_builder_.GetLeftBuffer(node_in_set, range.begin(), range.end());
-  uint32_t* p_right = partition_builder_.GetRightBuffer(node_in_set, range.begin(), range.end());
+  size_t* p_left = partition_builder_.GetLeftBuffer(node_in_set, range.begin(), range.end());
+  size_t* p_right = partition_builder_.GetRightBuffer(node_in_set, range.begin(), range.end());
 
   const bst_uint fid = tree[nid].SplitIndex();
   const bool default_left = tree[nid].DefaultLeft();
